CAS-255: update SPDK to 20.x
This change moves us to a newer SPDK version. As as result of internal
changes, we had to make some adjustments internally.

+++ mayastor/src/bdev/nexus/nexus_module.rs
@@ -2,6 +2,7 @@ use std::{cell::UnsafeCell, ffi::CString};
 
 use once_cell::sync::{Lazy, OnceCell};
 use serde_json::json;
+
 use spdk_sys::{
     spdk_bdev_module,
     spdk_bdev_module_examine_done,
@@ -11,7 +12,6 @@ use spdk_sys::{
     spdk_json_write_val_raw,
 };
 
-use super::instances;
 use crate::{
     bdev::nexus::{
         nexus_bdev::{Nexus, NexusState},
@@ -19,6 +19,8 @@ use crate::{
     },
     core::{Bdev, Reactor},
 };
+
+use super::instances;
 
 const NEXUS_NAME: &str = "NEXUS_CAS_MODULE";
 
@@ -181,7 +183,7 @@ impl NexusModule {
                 spdk_json_write_val_raw(
                     w,
                     data.as_ptr() as *const _,
-                    data.as_bytes().len(),
+                    data.as_bytes().len() as u64,
                 );
             }
         });
